Added functionality to MenuKaart

+++ app/src/main/java/com/example/appoftheyear/classLibrary/Dessert.java
@@ -1,7 +1,7 @@
 package com.example.appoftheyear.classLibrary;
 
 public class Dessert extends MenuItem {
-    public Dessert(String naam, int prijs) {
+    public Dessert(String naam, float prijs) {
         super(naam, prijs);
     }
 }
