+++ app/src/main/java/com/example/appoftheyear/classLibrary/Drink.java
@@ -1,7 +1,7 @@
 package com.example.appoftheyear.classLibrary;
 
 public class Drink extends MenuItem {
-    public Drink(String naam, int prijs) {
+    public Drink(String naam, float prijs) {
         super(naam, prijs);
     }
 }
