+++ app/src/main/java/com/example/appoftheyear/classLibrary/Hoofdgerecht.java
@@ -1,7 +1,7 @@
 package com.example.appoftheyear.classLibrary;
 
 public class Hoofdgerecht extends MenuItem{
-    public Hoofdgerecht(String naam, int prijs) {
+    public Hoofdgerecht(String naam, float prijs) {
         super(naam, prijs);
     }
 }
